Remove noQuery: let's use jQuery for now
I'm uncomfortable with the dependencies that are
in this app that use jQuery:

* Dialog box
* Selection behavior
* Ajax() behavior

All of these are core to the app and can be
performed from within the context of Chrome's
very-capable JS engine.  They map to:

* confirm()
* document.querySelector*
* Promise()

I'd like to continue this work to remove the
jQuery links.  It feels like an unnecessary tangle
to keep jQuery in here.